+++ bpm-engine-impl/src/main/java/io/takari/bpm/actions/FindAndCallActivityAction.java
@@ -17,7 +17,7 @@ public class FindAndCallActivityAction implements Action {
     @Override
     public String toString() {
         return "FindAndCallActivityAction[" +
-                "calledElement='" + calledElement + '\'' +
+                "calledElement=" + calledElement +
                 ']';
     }
 }