+++ bpm-engine-impl/src/main/java/io/takari/bpm/actions/FireOnScopeDestroyedInterceptorsAction.java
@@ -1,0 +1,23 @@
+package io.takari.bpm.actions;
+
+import java.util.UUID;
+
+public class FireOnScopeDestroyedInterceptorsAction implements Action {
+
+    private final UUID scopeId;
+
+    public FireOnScopeDestroyedInterceptorsAction(UUID scopeId) {
+        this.scopeId = scopeId;
+    }
+
+    public UUID getScopeId() {
+        return scopeId;
+    }
+
+    @Override
+    public String toString() {
+        return "FireOnScopeDestroyedInterceptorsAction[" +
+                "scopeId=" + scopeId +
+                ']';
+    }
+}