+++ bpm-engine-impl/src/main/java/io/takari/bpm/actions/PushScopeAction.java
@@ -6,16 +6,28 @@ import java.util.Arrays;
 
 public class PushScopeAction implements Action {
 
+    private final String definitionId;
+    private final String elementId;
     private final boolean exclusive;
     private final Command[] finishers;
 
-    public PushScopeAction(boolean exclusive) {
-        this(exclusive, (Command[])null);
+    public PushScopeAction(String definitionId, String elementId, boolean exclusive) {
+        this(definitionId, elementId, exclusive, (Command[])null);
     }
 
-    public PushScopeAction(boolean exclusive, Command... finishers) {
+    public PushScopeAction(String definitionId, String elementId, boolean exclusive, Command... finishers) {
+        this.definitionId = definitionId;
+        this.elementId = elementId;
         this.exclusive = exclusive;
         this.finishers = finishers;
+    }
+
+    public String getDefinitionId() {
+        return definitionId;
+    }
+
+    public String getElementId() {
+        return elementId;
     }
 
     public boolean isExclusive() {
@@ -28,9 +40,11 @@ public class PushScopeAction implements Action {
 
     @Override
     public String toString() {
-        return "PushScopeAction{" +
-                "exclusive=" + exclusive +
+        return "PushScopeAction[" +
+                "definitionId=" + definitionId +
+                ", elementId=" + elementId +
+                ", exclusive=" + exclusive +
                 ", finishers=" + Arrays.toString(finishers) +
-                '}';
+                ']';
     }
 }