+++ bpm-engine-impl/src/main/java/io/takari/bpm/elements/CallActivityHandler.java
@@ -59,7 +59,9 @@ public class CallActivityHandler implements ElementHandler {
 
         // push the new scope when the call begins
         // the scope will be "popped" by one of the finishing actions
-        actions.add(new PushCommandAction(new PerformActionsCommand(new PushScopeAction(false, finishers.toArray(new Command[finishers.size()])))));
+        actions.add(new PushCommandAction(new PerformActionsCommand(
+                new PushScopeAction(cmd.getDefinitionId(), cmd.getElementId(), false,
+                        finishers.toArray(new Command[finishers.size()])))));
 
         return actions;
     }